jaguar1: split Jaguar1 driver into src/jaguar1/ + decouple from jaguar3
Mirror the src/jaguar3/ layout: move the 12 Jaguar1-only module pairs
(RtlJaguarDevice, HalModule, RadioManagementModule, EepromManager,
FirmwareManager, Iqk8812a, Iqk8814a, PowerTracking8812a, PhydmWatchdog,
PhyTableLoader, FrameParser, BbDbgportReader) into src/jaguar1/.

Decouple the two generations — neither #includes the other's headers now
(only the WiFiDriver factory references both). The shared symbols jaguar3 used
from Jaguar1 headers are extracted into neutral headers in src/:
- RateDefinitions.h  (enum MGN_RATE)
- RxPacket.h         (Packet / rx_pkt_attrib / RX_PACKET_TYPE)
- TxDescBits.h       (SET_BITS_TO_LE_4BYTE + LE typedefs/helpers)

Make the shared RtlUsbAdapter generation-agnostic: drop infinite_read() (which
built the Jaguar1 FrameParser) — the RX loop now reads via the existing
generation-neutral bulk_read_raw() and parses in RtlJaguarDevice::read_frames()
(Jaguar3 parses via FrameParserJaguar3). No behavior change.

Build green all targets; ctest passes; grep confirms no cross-generation
#includes.
