+++ demo/main.cpp
@@ -10,8 +10,8 @@
 
 #include <libusb.h>
 
-#include "FrameParser.h"
-#include "RtlJaguarDevice.h"
+#include "jaguar1/FrameParser.h"
+#include "jaguar1/RtlJaguarDevice.h"
 #include "RtlUsbAdapter.h"
 #include "SignalStop.h"
 #include "WiFiDriver.h"
