+++ src/jaguar1/EepromManager.cpp
