+++ src/jaguar1/EepromManager.h
