+++ src/jaguar1/HalModule.cpp
