+++ src/jaguar1/HalModule.h
