+++ src/jaguar1/PowerTracking8812a.cpp
