+++ src/jaguar1/RadioManagementModule.h
@@ -48,94 +48,7 @@ enum RATE_SECTION {
   RATE_SECTION_NUM,
 };
 
-enum MGN_RATE {
-  MGN_1M = 0x02,
-  MGN_2M = 0x04,
-  MGN_5_5M = 0x0B,
-  MGN_6M = 0x0C,
-  MGN_9M = 0x12,
-  MGN_11M = 0x16,
-  MGN_12M = 0x18,
-  MGN_18M = 0x24,
-  MGN_24M = 0x30,
-  MGN_36M = 0x48,
-  MGN_48M = 0x60,
-  MGN_54M = 0x6C,
-  MGN_MCS32 = 0x7F,
-  MGN_MCS0,
-  MGN_MCS1,
-  MGN_MCS2,
-  MGN_MCS3,
-  MGN_MCS4,
-  MGN_MCS5,
-  MGN_MCS6,
-  MGN_MCS7,
-  MGN_MCS8,
-  MGN_MCS9,
-  MGN_MCS10,
-  MGN_MCS11,
-  MGN_MCS12,
-  MGN_MCS13,
-  MGN_MCS14,
-  MGN_MCS15,
-  MGN_MCS16,
-  MGN_MCS17,
-  MGN_MCS18,
-  MGN_MCS19,
-  MGN_MCS20,
-  MGN_MCS21,
-  MGN_MCS22,
-  MGN_MCS23,
-  MGN_MCS24,
-  MGN_MCS25,
-  MGN_MCS26,
-  MGN_MCS27,
-  MGN_MCS28,
-  MGN_MCS29,
-  MGN_MCS30,
-  MGN_MCS31,
-  MGN_VHT1SS_MCS0,
-  MGN_VHT1SS_MCS1,
-  MGN_VHT1SS_MCS2,
-  MGN_VHT1SS_MCS3,
-  MGN_VHT1SS_MCS4,
-  MGN_VHT1SS_MCS5,
-  MGN_VHT1SS_MCS6,
-  MGN_VHT1SS_MCS7,
-  MGN_VHT1SS_MCS8,
-  MGN_VHT1SS_MCS9,
-  MGN_VHT2SS_MCS0,
-  MGN_VHT2SS_MCS1,
-  MGN_VHT2SS_MCS2,
-  MGN_VHT2SS_MCS3,
-  MGN_VHT2SS_MCS4,
-  MGN_VHT2SS_MCS5,
-  MGN_VHT2SS_MCS6,
-  MGN_VHT2SS_MCS7,
-  MGN_VHT2SS_MCS8,
-  MGN_VHT2SS_MCS9,
-  MGN_VHT3SS_MCS0,
-  MGN_VHT3SS_MCS1,
-  MGN_VHT3SS_MCS2,
-  MGN_VHT3SS_MCS3,
-  MGN_VHT3SS_MCS4,
-  MGN_VHT3SS_MCS5,
-  MGN_VHT3SS_MCS6,
-  MGN_VHT3SS_MCS7,
-  MGN_VHT3SS_MCS8,
-  MGN_VHT3SS_MCS9,
-  MGN_VHT4SS_MCS0,
-  MGN_VHT4SS_MCS1,
-  MGN_VHT4SS_MCS2,
-  MGN_VHT4SS_MCS3,
-  MGN_VHT4SS_MCS4,
-  MGN_VHT4SS_MCS5,
-  MGN_VHT4SS_MCS6,
-  MGN_VHT4SS_MCS7,
-  MGN_VHT4SS_MCS8,
-  MGN_VHT4SS_MCS9,
-  MGN_UNKNOWN
-};
+#include "RateDefinitions.h"
 
 /* Read-only snapshot of the chip's thermal meter. `raw` is the live
  * RF[A][0x42][15:10] reading (0..63, Realtek "thermal units" — roughly
