+++ src/jaguar1/RtlJaguarDevice.cpp
@@ -375,6 +375,27 @@ bool RtlJaguarDevice::send_packet(const uint8_t *packet, size_t length) {
   return resp;
 }
 
+std::vector<Packet> RtlJaguarDevice::read_frames() {
+  /* Cover one full chip-side RX aggregate (the 8814 pairs a 20K DMA-agg
+   * threshold with a 32K host read). 32K buffer, an exact multiple of
+   * wMaxPacketSize so no short packet truncates the transfer. */
+  static constexpr int BUF_SIZE = 32 * 1024;
+  uint8_t buffer[BUF_SIZE] = {};
+  int n = _device.bulk_read_raw(buffer, sizeof(buffer), USB_TIMEOUT * 10);
+  if (n < 0) {
+    /* Rate-limit the error log + sleep so a fast-failing rc (e.g. NO_DEVICE
+     * after the chip drops off USB) can't spin the RX loop at full CPU. */
+    static uint64_t err_count = 0;
+    if ((err_count++ % 100) == 0)
+      _logger->error("bulk_read_raw failed with error: {} (count={})", n,
+                     err_count);
+    std::this_thread::sleep_for(std::chrono::milliseconds(50));
+    n = 0;
+  }
+  FrameParser fp{_logger};
+  return fp.recvbuf2recvframe(std::span<uint8_t>{buffer, (size_t)n});
+}
+
 void RtlJaguarDevice::Init(Action_ParsedRadioPacket packetProcessor,
                           SelectedChannel channel) {
   _packetProcessor = packetProcessor;
@@ -389,7 +410,7 @@ void RtlJaguarDevice::Init(Action_ParsedRadioPacket packetProcessor,
    * transfer there is a window between transfers where NO URB is posted; the
    * 8814's RX-DMA ring pauses in that gap and, under sparse traffic,
    * intermittently wedges after a short burst — the "RX stalls at ~10 frames"
-   * bug (8812/8821 tolerate it, the 8814 does not). infinite_read() is
+   * bug (8812/8821 tolerate it, the 8814 does not). read_frames() is
    * reentrant (local buffer + local FrameParser; libusb is thread-safe), so a
    * small worker pool restores the always-posted behaviour. Frame order across
    * workers is not preserved, which is fine for monitor-mode RX. Set
@@ -404,7 +425,7 @@ void RtlJaguarDevice::Init(Action_ParsedRadioPacket packetProcessor,
   for (int i = 0; i < rx_workers; ++i) {
     workers.emplace_back([this, &proc_mu]() {
       while (!should_stop && !g_devourer_should_stop) {
-        auto packets = _device.infinite_read();
+        auto packets = read_frames();
         if (packets.empty())
           continue;
         std::lock_guard<std::mutex> lk(proc_mu);
